server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "A geospatial analysis project leveraging advanced Python libraries."
authors = [
    { name = "Omdena" }
]
readme = "README.md"
requires-python = ">=3.12"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
asf_search = "8.0.1"
dask = "2024.12.1"
dask-expr = "1.1.21"
GDAL = "3.8.4"
geopandas = "1.0.1"
h5netcdf = "1.4.1"
h5py = "3.12.1"
matplotlib = "3.10.0"
numpy = "1.26.4"
pandas = "2.2.3"
pygmtsar = "2024.8.30.post5"
python-snappy = "0.7.3"
psutil = "6.1.1"
rasterio = "1.4.3"
scikit-image = "0.25.0"
scikit-learn = "1.6.0"
shapely = "2.0.6"
snappy = "3.1.1"
tqdm = "4.67.1"
vtk = "9.3.1"
xarray = "2024.11.0"
boto3 = "1.35.92"
pyvista = "0.44.2"
python-dotenv = "1.0.1"
xvfbwrapper = "0.2.9"
fastapi = {extras = ["all"], version = "^0.115.7"}
uvicorn = "^0.34.0"
pydantic = "^2.10.5"
pytest = "^8.3.4"
httpx = "^0.28.1"
rioxarray = "^0.18.2"
geojson = "^3.2.0"
sqlmodel = "^0.0.22"
sqlalchemy = "^2.0.37"
alembic = "^1.14.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "3.2.0"
netcdf4 = "^1.7.2"
moto = "^5.0.27"
osmnx = "^2.0.1"
pytest-mock = "^3.14.0"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.3"
black = "^24.10.0"
pre-commit = "^4.1.0"
